requirements.txt
absl-py==0.14.0
APScheduler==3.6.3
arabic-reshaper==2.1.3
asgiref==3.4.1
attrs==21.2.0
backports.entry-points-selectable==1.1.0
beautifulsoup4==4.10.0
cachetools==4.2.2
certifi==2021.5.30
cffi==1.15.0
chardet==3.0.4
charset-normalizer==2.0.6
clarifai==2.6.2
colorama==0.4.4
configparser==3.8.1
crayons==0.4.0
cryptography==35.0.0
cycler==0.10.0
defusedxml==0.7.1
distlib==0.3.3
dj-database-url==0.5.0
Django==3.2.8
django-allauth==0.45.0
django-countries==7.2.1
django-crispy-forms==1.13.0
django-filter==21.1
django-heroku==0.3.1
django-js-asset==1.2.2
django-mptt==0.13.4
django-mptt-autocomplete==0.1.3
django-paypal==1.1.1
django-phonenumber-field==5.2.0
django-qr-code==2.2.0
django-social-share==2.2.1
djangorestframework==3.12.4
djongo==1.3.6
EasyProcess==0.3
emoji==1.5.0
filelock==3.3.1
future==0.18.2
googleapis-common-protos==1.53.0
googletrans==3.0.0
grpcio==1.40.0
gunicorn==20.1.0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2021.9.1
html5lib==1.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
instapy==0.6.13
iso8601==0.1.16
jsonschema==2.6.0
kiwisolver==1.3.2
Markdown==3.3.4
matplotlib==3.4.3
MeaningCloud-python==2.0.0
mediapipe==0.8.7.3
numpy==1.21.2
oauthlib==3.1.1
opencv-contrib-python==4.5.3.56
opencv-python==4.5.3.56
paypal-checkout-serversdk==1.0.1
paypalhttp==1.0.1
pdfkit==0.6.1
pexpect==4.8.0
phonenumbers==8.12.36
Pillow==8.3.2
platformdirs==2.4.0
plyer==2.0.0
protobuf==3.18.0
psycopg2==2.9.1
ptyprocess==0.7.0
pycparser==2.20
pyFirmata==1.1.0
PyJWT==2.3.0
pymongo==3.12.0
pyOpenSSL==21.0.0
pyparsing==2.4.7
PyPDF2==1.26.0
pyserial==3.5
python-bidi==0.4.2
python-dateutil==2.8.2
python-decouple==3.4
python-telegram-bot==13.7
python3-openid==3.2.0
pytz==2021.1
PyYAML==5.4.1
regex==2021.8.28
reportlab==3.6.2
requests==2.26.0
requests-oauthlib==1.3.0
rfc3986==1.5.0
segno==1.3.3
selenium==3.141.0
serial==0.0.97
six==1.16.0
sniffio==1.2.0
social-auth-app-django==5.0.0
social-auth-core==4.1.0
soupsieve==2.2.1
sqlparse==0.2.4
Strip==0.0.1.dev14
stripe==2.61.0
tornado==6.1
tqdm==4.62.3
tzdata==2021.1
tzlocal==3.0
urllib3==1.26.7
virtualenv==20.8.1
virtualenvwrapper-win==1.2.6
webdriver-manager==3.4.2
webdriverdownloader==1.1.0.3
webencodings==0.5.1
whitenoise==5.3.0
xhtml2pdf==0.2.5
xls==0.0.2
XlsxWriter==3.0.1
